requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
openai>=1.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
